poetroid_device/requirements.txt
pillow
requests
opencv-python
pyyaml